refactor: Extract Cargo Shorts-related functions
Extract Cargo Cultist Shorts-related functions to
`kolmafia-cargo-cultist-shorts.d.ts`
