+++ src/index.d.ts
@@ -1,3 +1,4 @@
 import './global';
 export * from './kolmafia';
+export * from './kolmafia-cargo-cultist-shorts';
 export * from './kolmafia-modifier';
